+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes from v0.1 to v0.2
 API Changes
 -----------
 - Added executable SHE_CTE_SimulateImages
+- Added executable SHE_CTE_MeasureStatistics
 
 
 Dependency Changes